Moved PseudospectralTimeseriesOutputComp to common/TimeseriesOutputComp. It is now used in all transcriptions. Added a comparison dunder method __eq__ to GridData so that we can quickly test if two grid data objects have the same data even if they are separate instantiations. Some minor cleanup was necessary in SolveIVPTimeseriesOutputComp even though all SolveIVP machinery is deprecated through not yet removed until dymos 2.0.
